requirements.txt
requests>=2.28.1
tabulate>=0.9.0
python-dotenv>=0.20.0
beautifulsoup4>=4.11.1
matplotlib>=3.5.0
# PDF ve grafik oluşturma için gereken paketler
reportlab>=4.0.0
pdfrw>=0.4 